pyproject.toml
[tool.poetry]
name = "omelet-data-gen-python"
version = "0.1.0"
description = ""
authors = ["LarryKwon <65128957+LarryKwon@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
load-dotenv = "^0.1.0"
sqlalchemy = "^2.0.32"
openpyxl = "^3.1.5"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
aiohttp = "^3.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
